Moved remaining tests out of aggregator.
