+++ cps/src/test/java/org/jenkinsci/plugins/workflow/WorkflowTest.java
@@ -25,13 +25,26 @@
 package org.jenkinsci.plugins.workflow;
 
 import com.google.common.base.Function;
+import hudson.EnvVars;
+import hudson.model.Computer;
+import hudson.model.Descriptor;
 import hudson.model.Executor;
+import hudson.model.Node;
 import hudson.model.Queue;
+import hudson.model.Slave;
 import hudson.model.TaskListener;
 import hudson.model.User;
+import hudson.slaves.CommandLauncher;
+import hudson.slaves.NodeProperty;
+import hudson.slaves.RetentionStrategy;
+import hudson.slaves.SlaveComputer;
+import java.io.File;
+import java.io.IOException;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
+import javax.annotation.CheckForNull;
+import javax.annotation.Nonnull;
 import jenkins.model.Jenkins;
 import jenkins.security.QueueItemAuthenticatorConfiguration;
 import org.jenkinsci.plugins.scriptsecurity.scripts.ScriptApproval;
@@ -252,7 +265,7 @@ public class WorkflowTest extends SingleJobTestBase {
                 Map<String,String> slaveEnv = new HashMap<String,String>();
                 slaveEnv.put("BUILD_TAG", null);
                 slaveEnv.put("PERMACHINE", "set");
-                JenkinsRuleExt.createSpecialEnvSlave(story.j, "slave", null, slaveEnv);
+                createSpecialEnvSlave(story.j, "slave", null, slaveEnv);
                 p = jenkins().createProject(WorkflowJob.class, "demo");
                 p.setDefinition(new CpsFlowDefinition("node('slave') {\n"
                         + "  sh 'echo tag=$BUILD_TAG PERMACHINE=$PERMACHINE'\n"
@@ -289,6 +302,42 @@ public class WorkflowTest extends SingleJobTestBase {
                 assertNotNull(a.getEnvironment().get("PATH"));
             }
         });
+    }
+
+    // TODO add to jenkins-test-harness
+    /**
+     * Akin to {@link JenkinsRule#createSlave(String, String, EnvVars)} but allows {@link Computer#getEnvironment} to be controlled rather than directly modifying launchers.
+     * @param env variables to override in {@link Computer#getEnvironment}; null values will get unset even if defined in the test environment
+     * @see <a href="https://github.com/jenkinsci/jenkins/pull/1553/files#r23784822">explanation in core PR 1553</a>
+     */
+    public static Slave createSpecialEnvSlave(JenkinsRule rule, String nodeName, @CheckForNull String labels, Map<String,String> env) throws Exception {
+        @SuppressWarnings("deprecation") // keep consistency with original signature rather than force the caller to pass in a TemporaryFolder rule
+        File remoteFS = rule.createTmpDir();
+        SpecialEnvSlave slave = new SpecialEnvSlave(remoteFS, rule.createComputerLauncher(/* yes null */null), nodeName, labels != null ? labels : "", env);
+        rule.jenkins.addNode(slave);
+        return slave;
+    }
+    private static class SpecialEnvSlave extends Slave {
+        private final Map<String,String> env;
+        SpecialEnvSlave(File remoteFS, CommandLauncher launcher, String nodeName, @Nonnull String labels, Map<String,String> env) throws Descriptor.FormException, IOException {
+            super(nodeName, nodeName, remoteFS.getAbsolutePath(), 1, Node.Mode.NORMAL, labels, launcher, RetentionStrategy.NOOP, Collections.<NodeProperty<?>>emptyList());
+            this.env = env;
+        }
+        @Override public Computer createComputer() {
+            return new SpecialEnvComputer(this, env);
+        }
+    }
+    private static class SpecialEnvComputer extends SlaveComputer {
+        private final Map<String,String> env;
+        SpecialEnvComputer(SpecialEnvSlave slave, Map<String,String> env) {
+            super(slave);
+            this.env = env;
+        }
+        @Override public EnvVars getEnvironment() throws IOException, InterruptedException {
+            EnvVars env2 = super.getEnvironment();
+            env2.overrideAll(env);
+            return env2;
+        }
     }
 
 }
